Functioning version with unknown bug
Fixed the flux density issue, so absolute units of spectrum are
sound. However, the is a bug which appears to create spread in
the spectrum. Unknown origin. See evernote notes.

This commit is prior to making all floats double precision in F90
and compiling DISORT and f77 code as 8-byte real.