Functioning code for Pleiades run of 5gas+temp
This is the version that is running on pleiades for testing of 5gas
retrieval and T-P with nuisance parameters.

We now have:

- seperate PBS scripts for Herts and NASA
- Separate Makefiles for NASA and UH

Linelists are stored in /nobackup in brewster.py
THIS NEEDS TO BE MADE SIMPLE TO CHANGE. Perhaps an argument?

Perhaps an argument can set both Makefile version and Linelist location?

These can be kicked down the line. Next tasks:

- finalise dust coding - log(n_cond / n_gas)