Functioning fast code with disort option
This works with the new fast RT, and has disort option as an argument to
the python module

Build and makefile have been updated to compile this correctly.

Also fixed disagreement in the w1,w2 for the retrieval vs the
5gas_spectrum.dat

This branch should become the main branch.

Next to do:

Sort out the dust coding: profile in log(n_cond / n_gas)

Run a test of disort cloud spec vs fast cloud spec