Bug fixes and McNugget:new for postprocessing
Fixed various bugs associated with NaNs from Linelists, small edits to
code in improve speed. forwardmodel runs in <0.5seconds now.

We can now retrieve 5 gases (at NASA), and get pretty good answers.

Introduced McNuggets.py: this is a post-processor that adds Teff (and in
the future other derived quantities) to the posterior distributions from
the emcee. It is parallel, so can be slotted in after brewster for a
run.

Current bug is NaN in lnprob which crashes the Temperature retrieval
test, and NaN's in the last few thousand rows of Teff from McNugget