Updated nuggets, and other bits, gives bimodal PDF
This is merged between the old version that gives a good no cloud result
and the version that gives a bimodal

It appears to be down to the low-g fit being better at the short
wavelength, and the high-g fit better at longer wavelengths.

Using fewer walkers sends it to find only one of the solutions

Using Mike's alkalis pushes it to low-g solution with 16 walkers per
dim

Using R&R's alkalis sends it to high-g

Using > 24 walkers per dim finds both.

Cloud is pushed deep in both cases. Ready to merge with rayleigh branch
and setup Mie-retrievals...