Rayleigh optimised and included
Re-jigged the gas opacity calculations to put all the layer loops in the
subroutunes and the subroutines in a single module. this seems to be
about as quick as it was before, but is now somewhat tidier

The rayleigh can be turned off easily if we want a bit of extra speed.