Thin cloud version. BUG fixes, Incl TPmod bug
Fixed a load of bugs in the profile stuff, now using logPs for the theta
variables in TP and cloud modules.

Fixed a seriously stupied bug in TPmod that was sending profiles to Zero
and layer 0. duh!

Also fixed bugs in clouds that were ruining the cloud profile.

This *should* be running now....