Some changes to run 2M0355, also edits to sizes
mostly changes to run 2M0355, but also some edits to sizes and clouds
for doing mie code to small radii.

rg and rsig are now in log-um.

But there is a bug, which sends likelihood to inf after a few runs. Is
present in both laptop version and on cluster.

Once it goes to infinity, even previously good thetas return inf.