refactored testkit with model_spec module and bug fixes
Restructured testkit to put making the model spectrum in one place for
easy plotting etc. This creates the model_spec  which does all the
crap for the model, this makes plotting with runargs way easier and
interfaces with e.g optimiser better (if that ever works).

also Fixed some bugs in cloud.py