Added lognormal size distribution and some bugfixes
Have made the lognormal size distribution functional,though not fully
tested. This is accessed by making do_clouds = 2.

do_clouds and cloudy are no longer "logical" in main and clouds_mod,
but are instead integer that tells cloudatlas to use hansen (1) or
lognormal (2)

Parameters are the same. a and b for hansen are applied as rg for
lognormal, and rsig is obtained as 1+(b*4).  i.e rsig lies in 1 - 5
range.

Also fixed some bugs:

testkit.py: added upper bound to prior for cloud_tau0, which had been
absent. Hadn't been an issue before, but had a few cases wandering off
to crazy taus. Also, it makes the posterior finite.. which is always a
good thing.

brewtools.py: fixed PY3 problems loading big pickles in MacOS.