Fixes #12 Added simple test code for forward model,and small changes
Added bits for a simple test script of forward model and some bits of
testkit stuff.

Run the check_brewster.py to run two test runs. One is G570D based
forward model, and the other is 2M2224 Mie cloud version.

Both produce a spectrum which is compared to the benchmark
spectra (also in the git repository), and reports if there are bigger
than 1% differences.  Run this whenever changes are made to forward
model, or python bits. If changes to output are expected, verify
they're right and make new benchmark spectra.

Also added option to brewtools.get_endchain to specify a path to the
results files and runargs. This is default to ./, but can be pointed
to e.g. external disk.