Fixed stupid bug in cloud.py that set albedo = 0
In late 2017, a typo in cloud.py set w0 = 0 for grey and power law
clouds.
This was missed because albedo was never constrained anyhow.. but,
this reduced scatter in P0 and tau_cl for decks and slabs
respectively.

This is now fixed.

Albedo value for nested sampling revised to 0 - total absorption, as
it makes more sense...