updates to Profile 7, new profile 77, change slab cloud params
Fixed the priors on profiles type 7, and added a new smoothed version
of profile type 7, as 77 in emcee version,  which uses the smoothing prior from the
Line+2015 profile

Also fixed prior on profile type 7 in nested sampling

Finally swapped cloud base and top for slab clouds in emcee so it
matches nested sampling. i.e. we retrieve the cloud base and the
height, rather than the top and then fiddle with it afterwards so we
plot the base.