introducing Molliere hybrid profile (7) and changes for G395H data (#50)
* Update to include Molliere hybrid profile, also small fixes

- Profile 7 added
- maxwave increased to allow for higher resolution
- nestkit updated to align with P7 in TPmod and testkit
- gaslists updated
- cloudlist updated

* updates to Profile 7, new profile 77, change slab cloud params

Fixed the priors on profiles type 7, and added a new smoothed version
of profile type 7, as 77 in emcee version,  which uses the smoothing prior from the
Line+2015 profile

Also fixed prior on profile type 7 in nested sampling

Finally swapped cloud base and top for slab clouds in emcee so it
matches nested sampling. i.e. we retrieve the cloud base and the
height, rather than the top and then fiddle with it afterwards so we
plot the base.

* working version with profile type 7 forces RC boudary P<=100bars

Also:

- slab cloud base now retrieved instead of cloud top - this removes
need to swapping things around in the plotting notebook

- cloud and gas photosphere pressures in gnostics output now tau = 2/3

- a few gases added